+++ internal/sctp/sctp.go
@@ -35,10 +35,6 @@ import (
 )
 
 const (
-	NGAPPPID uint32 = 60
-)
-
-const (
 	SolSCTP = 132
 
 	SCTPBindxAddAddr = 0x01
