+++ nas/eps/esm_modify_test.go
@@ -27,8 +27,8 @@ func TestModifyEPSBearerContextRequestRoundTrip(t *testing.T) {
 		t.Fatalf("ESM header = % x, want first three bytes %x %x %x", wire[:3], 5<<4|0x02, 0, byte(MsgModifyEPSBearerContextRequest))
 	}
 
-	if wire[3] != ieiProtocolConfigurationOptions {
-		t.Fatalf("PCO IEI = %#x, want %#x", wire[3], ieiProtocolConfigurationOptions)
+	if wire[3] != protocolConfigurationOptionsIEI {
+		t.Fatalf("PCO IEI = %#x, want %#x", wire[3], protocolConfigurationOptionsIEI)
 	}
 
 	got, err := ParseModifyEPSBearerContextRequest(wire)
@@ -48,7 +48,7 @@ func TestModifyEPSBearerContextRequestRoundTrip(t *testing.T) {
 func TestModifyEPSBearerContextRequestAPNAMBRRoundTrip(t *testing.T) {
 	const dlBps, ulBps = 200 * 1_000_000, 100 * 1_000_000
 
-	apnambr := EncodeAPNAMBR(dlBps, ulBps).Marshal()
+	apnambr := APNAMBRFromBitsPerSecond(dlBps, ulBps).Marshal()
 	pco := BuildProtocolConfigurationOptions([][]byte{{1, 1, 1, 1}}, 1500)
 
 	req := &ModifyEPSBearerContextRequest{
@@ -64,12 +64,12 @@ func TestModifyEPSBearerContextRequestAPNAMBRRoundTrip(t *testing.T) {
 	}
 
 	// TS 24.301 §8.3.18.1: APN-AMBR precedes PCO in message order.
-	if wire[3] != ieiAPNAMBR {
-		t.Fatalf("first optional IEI = %#x, want APN-AMBR %#x", wire[3], ieiAPNAMBR)
+	if wire[3] != apnAMBRIEI {
+		t.Fatalf("first optional IEI = %#x, want APN-AMBR %#x", wire[3], apnAMBRIEI)
 	}
 
 	apnLen := int(wire[4])
-	if pcoOff := 5 + apnLen; pcoOff >= len(wire) || wire[pcoOff] != ieiProtocolConfigurationOptions {
+	if pcoOff := 5 + apnLen; pcoOff >= len(wire) || wire[pcoOff] != protocolConfigurationOptionsIEI {
 		t.Fatalf("PCO IEI not found after APN-AMBR at offset %d (wire % x)", pcoOff, wire)
 	}
 
@@ -98,7 +98,7 @@ func TestModifyEPSBearerContextRequestAPNAMBRRoundTrip(t *testing.T) {
 
 func TestModifyEPSBearerContextRequestNewEPSQoSRoundTrip(t *testing.T) {
 	epsQoS := EPSQoS{QCI: 7}.Marshal()
-	apnambr := EncodeAPNAMBR(200*1_000_000, 100*1_000_000).Marshal()
+	apnambr := APNAMBRFromBitsPerSecond(200*1_000_000, 100*1_000_000).Marshal()
 
 	req := &ModifyEPSBearerContextRequest{
 		EPSBearerIdentity: 5,
@@ -112,12 +112,12 @@ func TestModifyEPSBearerContextRequestNewEPSQoSRoundTrip(t *testing.T) {
 	}
 
 	// TS 24.301 §8.3.18.1: New EPS QoS precedes APN-AMBR.
-	if wire[3] != ieiNewEPSQoS {
-		t.Fatalf("first optional IEI = %#x, want New EPS QoS %#x", wire[3], ieiNewEPSQoS)
+	if wire[3] != newEPSQoSIEI {
+		t.Fatalf("first optional IEI = %#x, want New EPS QoS %#x", wire[3], newEPSQoSIEI)
 	}
 
 	qosLen := int(wire[4])
-	if ambrOff := 5 + qosLen; ambrOff >= len(wire) || wire[ambrOff] != ieiAPNAMBR {
+	if ambrOff := 5 + qosLen; ambrOff >= len(wire) || wire[ambrOff] != apnAMBRIEI {
 		t.Fatalf("APN-AMBR IEI not found after New EPS QoS at offset %d (wire % x)", ambrOff, wire)
 	}
